CHANGE: Remove the `setLoggedIn` setter from lower level components.
This refactoring follows the design from the ["Lifiting State
Up"](https://reactjs.org/docs/lifting-state-up.html) philosophy:

The information, if the user should be logged in or not is held by the top level
<App> component. It is set, though, by some lower component (namely <Login>,
<Register> and <Logout>). These lower level components do not should not modify
the state of <App> directly with passed down `setLoggedIn()` hooks.

Rather, the <App> component should be _notified_ via a callback function,
whenever the user has logged in or out.

The first step is to remove the passing of the `setLoggedIn` hook down the
component (this commit). The `onLogin()` and `onLogout()` callbacks will come
in the next commit. For now there are only comment placeholders.