CHANGE: Add `onLogin(user)` and `onLogout()` callbacks
Now we add new callbacks to the <Login>, <Register>, <Logout> and <UserBar>
components. The <App> component will hand these callbacks down the component
chain. When the user is logged in, <App> will first set the user name and then
the flag that the user is correctly logged in.

Note that <UserBar> passes the `onLogin` callback to _both_ the <Login> _and_
<Register> component! The `onLogin` takes one string parameter, namely the
username which got logged in. The <Login> component will pass the input from the
login field to this callback, while the <Register> component will pass the text
from the `registerUsername` to this callback. This way, <App> will be notified
of the current user, regardless where the username is coming from.

Quite nice not?